requirements.txt
# This file is intentionally left blank.
hydra-core
hydra-colorlog
langchain-community 
langchainhub 
langgraph